pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-cleaning-env"
version = "1.0.0"
description = "An OpenEnv environment for training AI agents on real-world data cleaning tasks"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "models*"]
